+++ spec/hatchet/nltk_spec.rb
@@ -12,7 +12,6 @@ RSpec.describe 'NLTK corpora support' do
           remote: -----> Downloading NLTK corpora…
           remote: -----> Downloading NLTK packages: city_database stopwords
           remote: .*: RuntimeWarning: 'nltk.downloader' found in sys.modules after import of package 'nltk', but prior to execution of 'nltk.downloader'; this may result in unpredictable behaviour
-          remote:   warn\\(RuntimeWarning\\(msg\\)\\)
           remote: \\[nltk_data\\] Downloading package city_database to
           remote: \\[nltk_data\\]     /tmp/build_.*/.heroku/python/nltk_data...
           remote: \\[nltk_data\\]   Unzipping corpora/city_database.zip.