+++ spec/spec_helper.rb
@@ -12,7 +12,7 @@ LATEST_PYTHON_3_8 = '3.8.16'
 LATEST_PYTHON_3_9 = '3.9.16'
 LATEST_PYTHON_3_10 = '3.10.10'
 LATEST_PYTHON_3_11 = '3.11.2'
-DEFAULT_PYTHON_VERSION = LATEST_PYTHON_3_10
+DEFAULT_PYTHON_VERSION = LATEST_PYTHON_3_11
 
 # Work around the return value for `default_buildpack` changing after deploy:
 # https://github.com/heroku/hatchet/issues/180